+++ include/Eigen/src/SVD/JacobiSVD_MKL.h
@@ -77,12 +77,12 @@ JacobiSVD<Matrix<EIGTYPE, Dynamic, Dynamic, EIGCOLROW, Dynamic, Dynamic>, ColPiv
   return *this; \
 }
 
-EIGEN_MKL_SVD(float,   double,        double, d, ColMajor, LAPACK_COL_MAJOR)
+EIGEN_MKL_SVD(float,   float,        double, d, ColMajor, LAPACK_COL_MAJOR)
 EIGEN_MKL_SVD(float,    float,         float , s, ColMajor, LAPACK_COL_MAJOR)
 EIGEN_MKL_SVD(dcomplex, MKL_Complex16, float, z, ColMajor, LAPACK_COL_MAJOR)
 EIGEN_MKL_SVD(scomplex, MKL_Complex8,  float , c, ColMajor, LAPACK_COL_MAJOR)
 
-EIGEN_MKL_SVD(float,   double,        double, d, RowMajor, LAPACK_ROW_MAJOR)
+EIGEN_MKL_SVD(float,   float,        double, d, RowMajor, LAPACK_ROW_MAJOR)
 EIGEN_MKL_SVD(float,    float,         float , s, RowMajor, LAPACK_ROW_MAJOR)
 EIGEN_MKL_SVD(dcomplex, MKL_Complex16, float, z, RowMajor, LAPACK_ROW_MAJOR)
 EIGEN_MKL_SVD(scomplex, MKL_Complex8,  float , c, RowMajor, LAPACK_ROW_MAJOR)
