+++ src/csvm/csvm_centroid.cc
@@ -5,9 +5,9 @@
 using namespace std;
 using namespace csvm;
 
-double Centroid::getDistanceSq(Feature f){
+float Centroid::getDistanceSq(Feature f){
    unsigned int nDims = f.content.size();
-   double sum = 0.0;
+   float sum = 0.0;
    for(size_t dIdx = 0; dIdx < nDims; ++dIdx){
       sum += (content[dIdx] - f.content[dIdx]) * (content[dIdx] - f.content[dIdx]);
    }
@@ -15,9 +15,9 @@ double Centroid::getDistanceSq(Feature f){
 }
 
 
-double Centroid::getDistanceSq(Centroid c) {
+float Centroid::getDistanceSq(Centroid c) {
 	unsigned int nDims = c.content.size();
-	double sum = 0.0;
+	float sum = 0.0;
 	for (size_t dIdx = 0; dIdx < nDims; ++dIdx) {
 		sum += (content[dIdx] - c.content[dIdx]) * (content[dIdx] - c.content[dIdx]);
 	}
