+++ src/csvm/csvm_frequency_matrix.cc
@@ -38,7 +38,7 @@ void FreqMatrix::setZeros(){
  
 }
 
-void FreqMatrix::addCombo(int wordA, int wordB, double activationA, double activationB){
+void FreqMatrix::addCombo(int wordA, int wordB, float activationA, float activationB){
    triangle[getTriangleIndex(wordA, wordB)] += activationA * activationB;
    totalInput += activationA * activationB;
 }
@@ -49,7 +49,7 @@ void FreqMatrix::normalize(){
    }
 }
 
-double FreqMatrix::getCombo(int wordA, int wordB){
+float FreqMatrix::getCombo(int wordA, int wordB){
    return triangle[getTriangleIndex(wordA, wordB)];
 }
